requirements.txt
google-adk>=1.0.0
google-cloud-firestore>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
notion-client>=2.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
